requirements.txt
MouseInfo==0.1.3
numpy==2.1.2
opencv-python==4.10.0.84
pillow==11.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0